Moved more macinery in the MVC sample
